Initial commit.
The kdl-org/kdl repository was added as a submodule to provide easy
access to the Test Cases, which are ran using HSpec. More extensive
property-based testing will be needed.

The parser is mostly complete but isn't completely faithful to the
SPEC just yet. Moreover, the formatter is comprised of basic Show
class instances; this should eventually be replaced by a pretty
printing library.

The structure of the code is currently a single file, and the API
is one function, `parseKDL` which doesn't yet report on the Parser's
errors.