Code refactoring: Types, Parser & Inernal modules.
The Types submodule contains data definition for a KDL document,
alongside their class instances.

Internal.hs contains parser definition that don't directly correspond
to nonterminal symbols in the grammar.

Parser.hs is the bulk of the code and houses the rest of the parsers.

Non-qualified imports from external modules are selective to avoid
namespace pollution. On the other hand internal module exports/imports
are not selective.